+++ graphsevalwindow.cpp
@@ -47,7 +47,6 @@ Drawer * GraphsEvalWindow::createNewDrawer(QString path)
     this->ui->totalC->setText(QString::number(this->drawer->proximity_curve_correlation));
     this->ui->totalI->setText(QString::number(this->drawer->proximity_curve_integral));
     this->ui->totalL->setText(QString::number(this->drawer->proximity_curve_local));
-    this->ui->totalS->setText(QString::number(this->drawer->proximity_curve_similarity));
     this->ui->totalPr->setText(QString::number(this->drawer->proximity_range));
     this->ui->totalPs->setText(QString::number(this->drawer->proximity_curve_shape));