+++ utils.h
@@ -1,0 +1,8 @@
+#ifndef UTILS_H
+#define UTILS_H
+
+#include <QStringList>
+
+QStringList scanDirItems(QString path, QString extension, QString relativePath);
+
+#endif // UTILS_H